Split headers to clean up include order
