+++ System/Proc/kern_cpu.cpp
@@ -1,5 +1,6 @@
 #include "kern_cpu.hpp"
 
+#include <vmm/pte.hpp>
 #include <vmm/vmm.hpp>
 #include <Boot/Boot.hpp>
 #include <kern_datatypes.hpp>
