+++ arch/aarch64/CMakeLists.txt
@@ -4,6 +4,7 @@ LIST(APPEND ARCH_HEADERS ints.hpp)
 LIST(APPEND ARCH_HEADERS mmu.hpp)
 LIST(APPEND ARCH_HEADERS sync.hpp)
 
+LIST(APPEND ARCH_HEADERS vmm/pte.hpp)
 LIST(APPEND ARCH_HEADERS vmm/vmm.hpp)
 LIST(APPEND ARCH_SOURCES vmm/vmm.cpp)
