+++ arch/aarch64/arch.hpp
@@ -1,17 +1,12 @@
-#ifndef _ARCH_HPP_
-#define _ARCH_HPP_
-
 #include <arch/aarch64/ints.hpp>
 #include <arch/aarch64/thread.hpp>
-#include <arch/aarch64/vmm/vmm.hpp>
+#include <arch/aarch64/cpureg.hpp>
 
 #include <arch/aarch64/sync.hpp>
 
-typedef struct PagingInfo
-{
-	UINT64 TTBR0;
-	UINT64 TTBR1;
-}PagingInfo;
+
+#ifndef _ARCH_HPP_
+#define _ARCH_HPP_
 
 namespace pantheon
 {
@@ -54,60 +49,6 @@ UINT64 ICOUNT();
 
 VOID PAUSE();
 VOID LIDT(void *IDT);
-
-}
-
-namespace CPUReg
-{
-
-FORCE_INLINE UINT64 R_TTBR0_EL1()
-{
-	UINT64 RetVal = 0;
-	asm volatile ("mrs %0, ttbr0_el1\n" : "=r"(RetVal) ::);
-	return RetVal;
-}
-
-FORCE_INLINE UINT64 R_TTBR1_EL1()
-{
-	UINT64 RetVal = 0;
-	asm volatile ("mrs %0, ttbr1_el1\n" : "=r"(RetVal) ::);
-	return RetVal;
-}
-
-FORCE_INLINE VOID W_TTBR0_EL1(UINT64 Val)
-{
-	asm volatile ("msr ttbr0_el1, %0\n" :: "r"(Val) :);
-}
-
-FORCE_INLINE VOID W_TTBR1_EL1(UINT64 Val)
-{
-	asm volatile ("msr ttbr1_el1, %0\n" :: "r"(Val) :);
-}
-
-FORCE_INLINE VOID W_MAIR_EL1(UINT64 Value)
-{
-	asm volatile ("msr mair_el1, %0\n" :: "r"(Value) :);
-}
-
-FORCE_INLINE UINT64 R_MAIR_EL1()
-{
-	UINT64 RetVal = 0;
-	asm volatile ("mrs %0, mair_el1\n" : "=r"(RetVal) ::);
-	return RetVal;
-}
-
-FORCE_INLINE VOID W_TCR_EL1(UINT64 Value)
-{
-	asm volatile ("msr tcr_el1, %0\n" :: "r"(Value) :);
-}
-
-FORCE_INLINE UINT64 R_TCR_EL1()
-{
-	UINT64 RetVal = 0;
-	asm volatile ("mrs %0, tcr_el1\n" : "=r"(RetVal) ::);
-	return RetVal;
-}
-
 
 }
